Refactor notification messsage to add link
