+++ packages/valist-web/forms/create-account.tsx
@@ -1,54 +1,54 @@
 import { z } from 'zod';
 import { ApolloCache } from '@apollo/client';
-import { ReleaseMeta, Client } from '@valist/sdk';
-import type { FileWithPath } from 'file-selector';
-import { Event } from 'ethers';
+import { AccountMeta, Client } from '@valist/sdk';
 import { handleEvent } from './events';
 import * as utils from './utils';
-import { versionRegex } from './common';
+import { shortnameRegex } from './common';
 
 export interface FormValues {
-  releaseName: string;
+  accountName: string;
   displayName: string;
+  website: string;
   description: string;
 }
 
 export const schema = z.object({
-  releaseName: z.string()
-    .min(3, { message: 'Release name should have at least 3 characters' })
-    .max(24, { message: 'Release name should not be longer than 24 characters' })
-    .regex(versionRegex, { message: 'Release name can only contain letters, numbers, and dashes' })
-    .refine((val: string) => val.toLocaleLowerCase() === val, { message: 'Release name can only contain lowercase letters' }),
+  accountName: z.string()
+    .min(3, { message: 'Account name should have at least 3 characters' })
+    .max(24, { message: 'Account name should not be longer than 24 characters' })
+    .regex(shortnameRegex, { message: 'Account name can only contain letters, numbers, and dashes' })
+    .refine((val) => val.toLocaleLowerCase() === val, { message: 'Account name can only contain lowercase letters' }),
   displayName: z.string()
     .min(3, { message: 'Display name should have at least 3 characters' })
     .max(24, { message: 'Display name should not be longer than 32 characters' }),
+  website: z.string(),
   description: z.string()
     .max(100, { message: 'Description should be shorter than 100 characters' }),
 });
 
-export async function createRelease(
+export async function createAccount(
   address: string | undefined,
-  projectId: string,
   image: File | undefined,
-  files: FileWithPath[],
+  members: string[],
   values: FormValues,
   valist: Client,
   cache: ApolloCache<any>,
 ): Promise<boolean | undefined> {
   try {
-  	utils.hideError();
+    utils.hideError();
 
     if (!address) {
       throw new Error('connect your wallet to continue');
     }
 
-    if (files.length === 0) {
-      throw new Error('files cannot be empty');
+    if (members.length === 0) {
+      throw new Error('members cannot be empty');
     }
 
-    const meta: ReleaseMeta = {
+    const meta: AccountMeta = {
       name: values.displayName,
       description: values.description,
+      external_url: values.website,
     };
 
     utils.showLoading('Uploading files');
@@ -56,14 +56,13 @@ export async function createRelease(
       meta.image = await valist.writeFile(image);
     }
 
-    meta.external_url = await valist.writeFolder(files);
-
     utils.updateLoading('Creating transaction');
-    const transaction = await valist.createRelease(projectId, values.releaseName, meta);
-
-    utils.updateLoading('Waiting for transaction', transaction.hash);
+    const transaction = await valist.createAccount(values.accountName, meta, members);
     const receipt = await transaction.wait();
-    receipt.events?.forEach((event: Event) => handleEvent(event, cache));
+
+    const message = <a href={`https://polygonscan.com/tx/${transaction.hash}`} rel="noreferrer" target="_blank">Waiting for transaction - View transaction</a>;
+    utils.updateLoading(message);
+    receipt.events?.forEach(event => handleEvent(event, cache));
 
     return true;
   } catch (error: any) {
