+++ packages/valist-web/forms/create-project.tsx
@@ -89,7 +89,8 @@ export async function createProject(
     utils.updateLoading('Creating transaction');
     const transaction = await valist.createProject(accountId, values.projectName, meta, members);
 
-    utils.updateLoading('Waiting for transaction', transaction.hash);
+    const message = <a href={`https://polygonscan.com/tx/${transaction.hash}`} rel="noreferrer" target="_blank">Waiting for transaction - View transaction</a>;
+    utils.updateLoading(message);
     const receipt = await transaction.wait();
     receipt.events?.forEach((event: Event) => handleEvent(event, cache));
