+++ packages/valist-web/forms/update-account.tsx
@@ -1,13 +1,30 @@
+import { z } from 'zod';
 import { ApolloCache } from '@apollo/client';
-import { Client } from '@valist/sdk';
+import { AccountMeta, Client } from '@valist/sdk';
 import { handleEvent } from './events';
 import * as utils from './utils';
+import { ExternalLink } from 'tabler-icons-react';
 
-export async function setProductRoyalty(
+export interface FormValues {
+  displayName: string;
+  website: string;
+  description: string;
+}
+
+export const schema = z.object({
+  displayName: z.string()
+    .min(3, { message: 'Display name should have at least 3 characters' })
+    .max(24, { message: 'Display name should not be longer than 32 characters' }),
+  website: z.string(),
+  description: z.string()
+    .max(100, { message: 'Description should be shorter than 100 characters' }),
+});
+
+export async function updateAccount(
   address: string | undefined,
-  projectId: string,
-  recipient: string,
-  amount: number,
+  accountId: string,
+  image: File | undefined,
+  values: FormValues,
   valist: Client,
   cache: ApolloCache<any>,
 ): Promise<boolean | undefined> {
@@ -18,8 +35,22 @@ export async function setProductRoyalty(
       throw new Error('connect your wallet to continue');
     }
 
-    utils.showLoading('Waiting for transaction');
-    const transaction = await valist.setProductRoyalty(projectId, recipient, amount);
+    const meta: AccountMeta = {
+      name: values.displayName,
+      description: values.description,
+      external_url: values.website,
+    };
+
+    utils.showLoading('Uploading files');
+    if (image) {
+      meta.image = await utils.writeFile(image, valist);
+    }
+
+    utils.updateLoading('Creating transaction');
+    const transaction = await valist.setAccountMeta(accountId, meta);
+
+    const message = <a href={`https://polygonscan.com/tx/${transaction.hash}`} rel="noreferrer" target="_blank">Waiting for transaction - View transaction</a>;
+    utils.updateLoading(message);
     const receipt = await transaction.wait();
     receipt.events?.forEach(event => handleEvent(event, cache));
 
@@ -32,10 +63,10 @@ export async function setProductRoyalty(
   }
 }
 
-export async function setProductLimit(
+export async function addAccountMember(
   address: string | undefined,
-  projectId: string,
-  limit: number,
+  accountId: string,
+  member: string,
   valist: Client,
   cache: ApolloCache<any>,
 ): Promise<boolean | undefined> {
@@ -46,10 +77,8 @@ export async function setProductLimit(
       throw new Error('connect your wallet to continue');
     }
 
-    utils.showLoading('Creating transaction');
-    const transaction = await valist.setProductLimit(projectId, limit);
-
-    utils.updateLoading('Waiting for transaction', transaction.hash);
+    utils.showLoading('Waiting for transaction');
+    const transaction = await valist.addAccountMember(accountId, member);
     const receipt = await transaction.wait();
     receipt.events?.forEach(event => handleEvent(event, cache));
 
@@ -62,10 +91,10 @@ export async function setProductLimit(
   }
 }
 
-export async function setProductPrice(
+export async function removeAccountMember(
   address: string | undefined,
-  projectId: string,
-  price: number,
+  accountId: string,
+  member: string,
   valist: Client,
   cache: ApolloCache<any>,
 ): Promise<boolean | undefined> {
@@ -77,35 +106,7 @@ export async function setProductPrice(
     }
 
     utils.showLoading('Waiting for transaction');
-    const transaction = await valist.setProductPrice(projectId, price);
-    const receipt = await transaction.wait();
-    receipt.events?.forEach(event => handleEvent(event, cache));
-
-    return true;
-  } catch (error: any) {
-    utils.showError(error);
-    console.log(error);
-  } finally {
-    utils.hideLoading();
-  }
-}
-
-export async function withdrawProductBalance(
-  address: string | undefined,
-  projectId: string,
-  recipient: string,
-  valist: Client,
-  cache: ApolloCache<any>,
-): Promise<boolean | undefined> {
-    try {
-    utils.hideError();
-
-    if (!address) {
-      throw new Error('connect your wallet to continue');
-    }
-
-    utils.showLoading('Waiting for transaction');
-    const transaction = await valist.withdrawProductBalance(projectId, recipient);
+    const transaction = await valist.removeAccountMember(accountId, member);
     const receipt = await transaction.wait();
     receipt.events?.forEach(event => handleEvent(event, cache));
