+++ packages/valist-web/forms/update-project.tsx
@@ -80,7 +80,8 @@ export async function updateProject(
     utils.updateLoading('Creating transaction');
     const transaction = await valist.setProjectMeta(projectId, meta);
 
-    utils.updateLoading('Waiting for transaction', transaction.hash);
+    const message = <a href={`https://polygonscan.com/tx/${transaction.hash}`} rel="noreferrer" target="_blank">Waiting for transaction - View transaction</a>;
+    utils.updateLoading(message);
     const receipt = await transaction.wait();
     receipt.events?.forEach(event => handleEvent(event, cache));
